* refactor(embedding.py): Remove GeneralTextEmbedding class from embedding.py * feat(speculative.py): Add GeneralTextEmbedding class to speculative.py
